Get upload expenses page working.
