+++ src/scripts/helpers/index.mjs
@@ -216,3 +216,35 @@ export const uid = () => localStorage.getItem('uid') || null;
 export const isPayingUser = () => !!localStorage.getItem('isPayingUser');
 
 export const token = () => localStorage.getItem('token') || null;
+
+// utility function for converting CSVs to array of objects
+// https://gist.github.com/plbowers/7560ae793613ee839151624182133159
+export const csvStringToArray = (strData, header = true) => {
+  const objPattern = new RegExp(("(\\,|\\r?\\n|\\r|^)(?:\"((?:\\\\.|\"\"|[^\\\\\"])*)\"|([^\\,\"\\r\\n]*))"),"gi");
+  let arrMatches = null;
+  let arrData = [[]];
+  while (arrMatches = objPattern.exec(strData)) {
+    if (arrMatches[1].length && arrMatches[1] !== ",") {
+      arrData.push([]);
+    }
+
+    arrData[arrData.length - 1].push(arrMatches[2]
+      ? arrMatches[2].replace(new RegExp( "[\\\\\"](.)", "g" ), '$1')
+      : arrMatches[3]);
+  }
+
+  if (header) {
+    let hData = arrData.shift();
+    let hashData = arrData.map(row => {
+      let i = 0;
+      return hData.reduce((acc, key) => {
+        acc[key] = row[i++];
+        return acc;
+      }, {});
+    });
+
+    return hashData;
+  }
+
+  return arrData;
+};
