+++ src/settings/index.njk
@@ -92,6 +92,10 @@
     formatting, but there are templates available to help with that process.
   </p>
   <p>
+    If you need help with formatting, you can contact us at
+    {% include "components/contact-email.njk" %}
+  </p>
+  <p>
     <a href="/settings/upload-expenses">
       Upload Expenses
     </a>
